\name{A3}
\alias{A3}
\title{Trajectory}
\description{Trajectory, locally stored, from envirocar.org, see example below how it was imported}
\usage{
data(A3)
}
\keyword{datasets}
\examples{
library(spacetime)
data(A3)
dim(A3)
# see demo(A3) to see how A3 was fetched, and created from the web service
}
