% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBirthsChecker.R
\name{addBirthsChecker}
\alias{addBirthsChecker}
\title{Title}
\usage{
addBirthsChecker(connection, trajectoryAnalysisArgs, trajectoryLocalArgs)
}
\arguments{
\item{trajectoryLocalArgs}{}
}
\value{

}
\description{
Title
}
