% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIgraph.R
\name{addNumcohortCustom}
\alias{addNumcohortCustom}
\title{Adds numcohortCustom value to graph edges - actual number of people (out of all people who have EVENTNAME) on that edge}
\usage{
addNumcohortCustom(g, eventname, actualTrajs)
}
\arguments{
\item{actualTrajs}{}
}
\value{

}
\description{
Adds numcohortCustom value to graph edges - actual number of people (out of all people who have EVENTNAME) on that edge
}
