% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignActualTrajectoriesToGraph.R
\name{alignActualTrajectoriesToGraph}
\alias{alignActualTrajectoriesToGraph}
\title{Adds numcohortCustom value to graph edges - actual number of people (out of all people who have EVENTNAME) on that edge}
\usage{
alignActualTrajectoriesToGraph(
  connection,
  trajectoryAnalysisArgs,
  trajectoryLocalArgs,
  g,
  eventname,
  limit = 1000,
  filename = file.path(getwd(), "trajectories.csv")
)
}
\arguments{
\item{connection}{Database connection object created by createConnectionDetails() method in DatabaseConnector package}

\item{trajectoryAnalysisArgs}{TrajectoryAnalysisArgs object that must be created by createTrajectoryAnalysisArgs() method}

\item{trajectoryLocalArgs}{TrajectoryLocalArgs object that must be created by createTrajectoryLocalArgs() method}

\item{g}{An igrpah object that is created by specific graph functions in this package}

\item{eventname}{Event name (concept name) through which the trajectories are analyzed}

\item{limit}{Max number of trajectories to align (to limit the analysis). Set to NA if no limit.}

\item{filename}{Full path tforoutput file for trajectory counts}
}
\value{

}
\description{
Adds numcohortCustom value to graph edges - actual number of people (out of all people who have EVENTNAME) on that edge
}
