% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignActualTrajectoriesToGraphFull.R
\name{alignActualTrajectoriesToGraphFull}
\alias{alignActualTrajectoriesToGraphFull}
\title{Adds numcohortCustom value to graph nodes and edges - actual number of people on that edge/node}
\usage{
alignActualTrajectoriesToGraphFull(
  connection,
  trajectoryAnalysisArgs,
  trajectoryLocalArgs,
  g
)
}
\arguments{
\item{connection}{Database connection object created by createConnectionDetails() method in DatabaseConnector package}

\item{trajectoryAnalysisArgs}{TrajectoryAnalysisArgs object that must be created by createTrajectoryAnalysisArgs() method}

\item{trajectoryLocalArgs}{TrajectoryLocalArgs object that must be created by createTrajectoryLocalArgs() method}

\item{g}{An igrpah object that is created by specific graph functions in this package}
}
\description{
Adds numcohortCustom value to graph nodes and edges - actual number of people on that edge/node
}
