% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{alignTrajectoriesToGraph}
\alias{alignTrajectoriesToGraph}
\title{Takes actual event trajectories of the cohort and puts them to the given graph. Counts all different trajectories. Expects that the graph is created by createAlignmentTables() method which has created necessary database tables}
\usage{
alignTrajectoriesToGraph(
  connection,
  trajectoryAnalysisArgs,
  trajectoryLocalArgs,
  g
)
}
\arguments{
\item{g}{TrajectoriesGraph object, that is created by createAlignmentTables() method (that ensures that necessary database tables exist)}
}
\value{

}
\description{
Takes actual event trajectories of the cohort and puts them to the given graph. Counts all different trajectories. Expects that the graph is created by createAlignmentTables() method which has created necessary database tables
}
