% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{createAlignmentTableNew}
\alias{createAlignmentTableNew}
\title{Creates and fills alignment tables to the database for a given graph.
Particularly, creates table "graph_events" for events and table "graph_event_pairs" for the pairs}
\usage{
createAlignmentTableNew(
  connection,
  trajectoryAnalysisArgs,
  trajectoryLocalArgs,
  g
)
}
\arguments{
\item{g}{TrajectoriesGraph object}
}
\value{
TrajectoriesGraph object with filled E(g)$alignedTrajsCount, E(g)$alignedTrajsProb and V(g)$alignedTrajsCount values
}
\description{
Creates and fills alignment tables to the database for a given graph.
Particularly, creates table "graph_events" for events and table "graph_event_pairs" for the pairs
}
