% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{createAlignmentTables}
\alias{createAlignmentTables}
\title{Creates and fills alignment tables to the database for a given graph.
Particularly, creates table "graph_events" for events and table "graph_event_pairs" for the pairs
Note that these pairs are not "between any combination" of events like in discovery/validation analysis,
but specifically between the events that are occurring temporally next to each other.
For example, A->B->C produces pairs A->B and B->C only.}
\usage{
createAlignmentTables(
  connection,
  trajectoryAnalysisArgs,
  trajectoryLocalArgs,
  g
)
}
\arguments{
\item{g}{TrajectoriesGraph object}
}
\value{
TrajectoriesGraph object with filled E(g)$alignedTrajsCount, E(g)$alignedTrajsProb and V(g)$alignedTrajsCount values
}
\description{
Also note that "graph_event_pairs" does not include pairs between events that occur on the same day.
In case B and C occur on the same day on A->B/C->D trajectory, the pairs A->B, A->C, A->D, B->D, C->D are created.
Therefore, a single eventperiod can still produce several "paths" in the graph from A to D.
}
\details{
In case a patient has A->B->C->D, but there are only edges A->B, A->D, C->D on the graph, it also produces 2 separate trajectories: A->B and C->D
}
