% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAndFillCohortTable.R
\name{createAndFillCohortTable}
\alias{createAndFillCohortTable}
\title{Creates and fills cohort table that is going to be used by the Trajectories package}
\usage{
createAndFillCohortTable(
  connection,
  trajectoryAnalysisArgs,
  trajectoryLocalArgs
)
}
\arguments{
\item{connection}{DatabaseConnectorConnection object that is used to connect with database}

\item{trajectoryAnalysisArgs}{Object created by Trajectories:::createTrajectoryAnalysisArgs() method}

\item{trajectoryLocalArgs}{Object created by Trajectories:::createTrajectoryLocalArgs() method}
}
\description{
Creates and fills cohort table that is going to be used by the Trajectories package
}
