% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCohortTable.R
\name{createCohortTable}
\alias{createCohortTable}
\title{Function to create empty cohort table. Script is based on  https://ohdsi.github.io/TheBookOfOhdsi/SqlAndR.html#implementing-the-study-using-sql-and-r (21th May 2020)}
\usage{
createCohortTable(
  packageName,
  connection,
  dbms,
  sqlRole = F,
  cohortTableSchema = cohortTableSchema,
  cohortTable = "my_cohorts"
)
}
\arguments{
\item{connection}{DatabaseConnectorConnection object that is used to connect with database}

\item{sqlRole}{SQL role that is used to create cohort table in cohortTableSchema. Set to FALSE if a specific role is not needed}

\item{cohortTable}{}
}
\value{

}
\description{
Function to create empty cohort table. Script is based on  https://ohdsi.github.io/TheBookOfOhdsi/SqlAndR.html#implementing-the-study-using-sql-and-r (21th May 2020)
}
