% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEventPairsTable.R
\name{createEventPairsTable}
\alias{createEventPairsTable}
\title{Creates event pairs table and populates it with the data}
\usage{
createEventPairsTable(
  connection,
  trajectoryAnalysisArgs,
  trajectoryLocalArgs,
  eventParametersFilename = "event_parameters.txt"
)
}
\arguments{
\item{connection}{DatabaseConnectorConnection object that is used to connect with database}

\item{trajectoryAnalysisArgs}{TrajectoryAnalysisArgs object that must be created by createTrajectoryAnalysisArgs() method}

\item{trajectoryLocalArgs}{TrajectoryLocalArgs object that must be created by createTrajectoryLocalArgs() method}

\item{eventParametersFilename}{Filename where the parameters of creating event pairs are written (so that one could later see what what where the exact parameters). Default value is 'event_parameters.txt'. If set to FALSE, no output file is created.}
}
\value{

}
\description{
Creates event pairs table and populates it with the data
}
