% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEventsTable.R
\name{createEventsTable}
\alias{createEventsTable}
\title{Function to create empty cohort table. Script is based on  https://ohdsi.github.io/TheBookOfOhdsi/SqlAndR.html#implementing-the-study-using-sql-and-r (21th May 2020)}
\usage{
createEventsTable(
  connection,
  trajectoryAnalysisArgs,
  trajectoryLocalArgs,
  baseCohort = list(`0` = "All persons"),
  targetCohorts = list(`701` = "PAP test 30-65", `700` = "Kolposkoopia", `672` =
    "HPV test", `699` = "Observation of ASCUS 30-59", `696` =
    "Observation of LSIL 30-59", `692` = "Observation of ASC-H 30-59", `690` =
    "Observation of HSIL 30-59", `654` = "Observation of NILM 30-59")
)
}
\arguments{
\item{connection}{DatabaseConnectorConnection object that is used to connect with database}

\item{trajectoryAnalysisArgs}{TrajectoryAnalysisArgs object that must be created by createTrajectoryAnalysisArgs() method}

\item{trajectoryLocalArgs}{TrajectoryLocalArgs object that must be created by createTrajectoryLocalArgs() method}
}
\value{

}
\description{
Function to create empty cohort table. Script is based on  https://ohdsi.github.io/TheBookOfOhdsi/SqlAndR.html#implementing-the-study-using-sql-and-r (21th May 2020)
}
