% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGraph.R
\name{createGraph}
\alias{createGraph}
\title{Builds an igraph object based on event pairs data in "eventPairResultsFilename" file}
\usage{
createGraph(eventPairResultsFilename)
}
\arguments{
\item{eventPairResultsFilename}{Full path to event pairs file}
}
\value{

}
\description{
Builds an igraph object based on event pairs data in "eventPairResultsFilename" file
}
