% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIgraph.R
\name{createIgraph}
\alias{createIgraph}
\title{Title}
\usage{
createIgraph(
  connection,
  trajectoryAnalysisArgs,
  trajectoryLocalArgs,
  eventPairResultsFilename,
  outputFolder,
  eventName = NA
)
}
\arguments{
\item{outputFolder}{Output folder (should exist) without trailing slash}

\item{eventName}{Exact concept name that is used for building trajectories. Must exist in event pairs data table. If not specified (NA) (recommended) creates trajectories for top 5 events.}

\item{sqlRole}{Database role that is used when creating tables to 'resultsSchema'. Set to F if a specific role is not needed.}
}
\value{

}
\description{
Title
}
