% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTrajectoriesGraph.R
\name{createTrajectoriesGraph}
\alias{createTrajectoriesGraph}
\title{Builds a TrajectoriesGraph object based on event pairs data in "eventPairResultsFilename" file}
\usage{
createTrajectoriesGraph(eventPairResultsFilename)
}
\arguments{
\item{eventPairResultsFilename}{Full path to event pairs file}
}
\value{

}
\description{
It is actually an igraph object but has some specific attributes (counts, colors) necessary for Trajectories package
}
\details{
Vertex attributes:
(ID=concept_name), concept_id, count, color, labelcolor

Edge attributes:
e1 (name), e1_concept_id, e2 (name), e2_concept_id, e1_count, effect, prob, numcohortExact. numcohortExact is a number of event periods that had E1->E2 as immediate order (no intermediate events).
}
