% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryArguments.R
\name{createTrajectoryLocalArgs}
\alias{createTrajectoryLocalArgs}
\title{Title}
\usage{
createTrajectoryLocalArgs(
  cdmDatabaseSchema,
  vocabDatabaseSchema,
  resultsSchema,
  oracleTempSchema,
  sqlRole,
  prefixForResultTableNames,
  cohortTableSchema,
  cohortTable,
  cohortId,
  mainOutputFolder,
  cohortSqlFile
)
}
\arguments{
\item{cdmDatabaseSchema}{schema containing source data}

\item{vocabDatabaseSchema}{schema containing concepts library}

\item{resultsSchema}{schema the user has writing access to (used ti write new tables into)}

\item{oracleTempSchema}{In case you are using oracle, schema for temporary tables need to be specified. A schema where temp tables can be created in Oracle. Otherwise leave it as it is (is not used)}

\item{sqlRole}{Role to use in SQL for writing tables in 'resultsSchema'. It should also have access to 'cdmDatabaseSchema' and 'vocabDatabaseSchema'. Set to FALSE (or F) if setting to a specific role is not needed. In Estonian data is has to be hwisc_epi_ohdsi_dev_create}

\item{prefixForResultTableNames}{To avoid any collision with output table names (when someone runs the same analysis in parallel) we use a prefix for all table names that consists of 2 letters from username and 2 random characters. In case there is no username given (e.g Eunomia package, the prefix is simply 2 random characters)
Cohort table specifications (currently in development. Leave them as they are)}

\item{mainOutputFolder}{# Change the output folder path. This is the folder where the final results are produced.Use full path and do NOT add trailing slash! The folder must already exist.}

\item{cohortSqlFile}{}
}
\value{

}
\description{
Title
}
