% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createValidationSet.R
\name{createValidationSet}
\alias{createValidationSet}
\title{Randomly assigns event pairs from the cohort to validation set by using the given ratio}
\usage{
createValidationSet(
  connection,
  trajectoryAnalysisArgs,
  trajectoryLocalArgs,
  size = 0.5
)
}
\arguments{
\item{connection}{DatabaseConnectorConnection object that is used to connect with database}

\item{trajectoryAnalysisArgs}{Object created by Trajectories::createTrajectoryAnalysisArgs() method}

\item{trajectoryLocalArgs}{Object created by Trajectories::createTrajectoryLocalArgs() method}

\item{size}{Ratio for assigning event pairs from the cohort to validation set.}
}
\value{

}
\description{
Randomly assigns event pairs from the cohort to validation set by using the given ratio
}
