% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createValidationSetup.R
\name{createValidationSetup}
\alias{createValidationSetup}
\title{Creats setup file for the validation of the results in a separate database}
\usage{
createValidationSetup(trajectoryAnalysisArgs, trajectoryLocalArgs)
}
\arguments{
\item{trajectoryAnalysisArgs}{Object created by Trajectories::createTrajectoryAnalysisArgs() method}

\item{trajectoryLocalArgs}{Object created by Trajectories::createTrajectoryLocalArgs() method}
}
\value{

}
\description{
Creats setup file for the validation of the results in a separate database
}
