\name{crop}
\alias{crop}
\title{
Crop an STTDF object using a SpatialPolygons object
}
\description{
Get all points from an STTDF object that overlay with a SpatialPolygons object and return as an STTDF object
}
\usage{
crop(sttdf, pol)
}

\arguments{
  \item{sttdf}{
  a STTDF object
}
  \item{pol}{
  a SpatialPolygons object
}
}

\value{
\code{results} return an STTDF object
}
\references{
TBA
}
\author{
Jinlong Yang: jinlong@psu.edu
}