% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{detectSubgraphs}
\alias{detectSubgraphs}
\title{Title}
\usage{
detectSubgraphs(all_trajs)
}
\arguments{
\item{all_trajs}{Must have columns trajectory.str, trajectory.count, length}
}
\value{

}
\description{
Title
}
