% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover.R
\name{discover}
\alias{discover}
\title{The main function to run the Trajectory package in discovery mode - that is, to identify the temporal event sequences in the cohort of your local OMOP CDM}
\usage{
discover(
  connection,
  trajectoryLocalArgs,
  createCohort = T,
  validationSetSize = 0.5,
  createEventPairsTable = T,
  runDiscoveryAnalysis = T,
  forceRecalculationOfAnalysis = F,
  createFilteredFullgraphs = T,
  createGraphsForSelectedEvents = T,
  selfValidate = F,
  cleanup = T
)
}
\arguments{
\item{connection}{Database connection object created by createConnectionDetails() method in DatabaseConnector package}

\item{trajectoryLocalArgs}{TrajectoryLocalArgs object that must be created by createTrajectoryLocalArgs() method}

\item{createCohort}{Builds a study cohort in the database. If validationSetSize is set to non-zero, splits the cohort into disovery and validation set.}

\item{validationSetSize}{If set to non-zero, splits the whole cohort into discovery and (self-)validation set. Is meaningful only if selfValidate=T. Allowed values are in range 0..1.}

\item{createEventPairsTable}{Builds all event pairs and necessary data tables in the database for the analysis. Also clears all results from the database if they exist.}

\item{runDiscoveryAnalysis}{Run the actual directionality analysis of all event pairs.}

\item{forceRecalculationOfAnalysis}{Forces deleting previous results from the database and rerunning the whole discovery analysis. Useful mostly in case something goes wrong and you need to force the recalculation (it is, when debugging). In normal circumstances using F is safe.}

\item{createFilteredFullgraphs}{Builds graphs based on the results.}

\item{createGraphsForSelectedEvents}{Builds graphs for selected events (event ID-s taken from trajectoryAnalysisArgs$eventIdsForGraphs).}

\item{selfValidate}{Normally, set to F/FALSE as it is always better to validate your results in another database. However, if you want to validate your results in your own database, then set selfValidate=T and validationSetSize=some meaningful proportion (for example, 0.5). In such case, the discovery analysis is actually conducted on half of the data and the results are then validated on another half.}

\item{cleanup}{Drops tables from the database that were created during various stages of the analysis.}
}
\value{

}
\description{
The main function to run the Trajectory package in discovery mode - that is, to identify the temporal event sequences in the cohort of your local OMOP CDM
}
