% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropCohortTable.R
\name{dropCohortTable}
\alias{dropCohortTable}
\title{Function to drop cohort table.}
\usage{
dropCohortTable(connection, trajectoryAnalysisArgs, trajectoryLocalArgs)
}
\arguments{
\item{connection}{DatabaseConnectorConnection object that is used to connect with database}

\item{sqlRole}{SQL role that is used to drop cohort table in cohortTableSchema. Set to FALSE if a specific role is not needed}

\item{cohortTable}{}
}
\value{

}
\description{
Function to drop cohort table.
}
