% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatedTimeRemaining.R
\name{estimatedTimeRemaining}
\alias{estimatedTimeRemaining}
\title{Returns estimated time remaining}
\usage{
estimatedTimeRemaining(progress_perc = 0.25, starttime = Sys.time())
}
\arguments{
\item{progress_perc}{The percentage of task progress}

\item{starttime}{Time when current task was started}
}
\value{

}
\description{
Returns estimated time remaining in a textual manner, based on the current progress percentage
}
