% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillCohortTable.R
\name{fillCohortTable}
\alias{fillCohortTable}
\title{Function to fill the cohort table.}
\usage{
fillCohortTable(connection, trajectoryAnalysisArgs, trajectoryLocalArgs)
}
\arguments{
\item{connection}{DatabaseConnectorConnection object that is used to connect with database}

\item{trajectoryAnalysisArgs}{TrajectoryAnalysisArgs object that must be created by createTrajectoryAnalysisArgs() method}

\item{trajectoryLocalArgs}{TrajectoryLocalArgs object that must be created by createTrajectoryLocalArgs() method}
}
\value{

}
\description{
Function to fill the cohort table.
}
