% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIgraph.R
\name{filterIgraphCrossingEvent}
\alias{filterIgraphCrossingEvent}
\title{Creates a subgraph of edges that go through EVETNAME. Starts building it from EVENTNAME, adds most probable edge. Then takes both events together, adds most probable edge out from these two etc. Limits to limitOfNodes nodes.}
\usage{
filterIgraphCrossingEvent(
  g,
  eventname = "clopidogrel",
  limitOfNodes = F,
  edge_param_to_sort_by = c("effect", "numcohortExact", "numcohortCustom",
    "effectCount", "prob")
)
}
\arguments{
\item{limitOfNodes}{}
}
\value{

}
\description{
Creates a subgraph of edges that go through EVETNAME. Starts building it from EVENTNAME, adds most probable edge. Then takes both events together, adds most probable edge out from these two etc. Limits to limitOfNodes nodes.
}
