% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIgraph.R
\name{filterIgraphRemoveLowEffectLinks}
\alias{filterIgraphRemoveLowEffectLinks}
\title{Title}
\usage{
filterIgraphRemoveLowEffectLinks(
  g,
  edge_param_to_sort_by = c("effect", "numcohortExact", "numcohortCustom")
)
}
\arguments{
\item{g}{}
}
\value{

}
\description{
Title
}
