% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/filterIgraphRemoveLowEffectLinksAndOrphanNodes.R
\name{filterIgraphRemoveLowEffectLinksAndOrphanNodes}
\alias{filterIgraphRemoveLowEffectLinksAndOrphanNodes}
\title{Title}
\usage{
filterIgraphRemoveLowEffectLinksAndOrphanNodes(
  g,
  limitOfLinks = 20,
  edge_param_to_sort_by = c("effect", "numcohortExact", "numcohortCustom",
    "effectCount", "prob")
)
}
\arguments{
\item{g}{igraph object}

\item{limitOfLinks}{limitOfLinks}

\item{edge_param_to_sort_by}{edge_param_to_sort_by}
}
\value{

}
\description{
Title
}
