% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTop5Events.R
\name{getTop5Events}
\alias{getTop5Events}
\title{Takes most prevalent top5 nodes from the graph (by count) and returns the corresponding node list}
\usage{
getTop5Events(g)
}
\arguments{
\item{g}{}
}
\value{

}
\description{
Takes most prevalent top5 nodes from the graph (by count) and returns the corresponding node list
}
