% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{InitLogger}
\alias{InitLogger}
\title{Initializes logger for Trajectories package}
\usage{
InitLogger(logfile, threshold = logger:::INFO)
}
\arguments{
\item{logfile}{Full path to log file}

\item{threshold}{log_threshold values from "logger" package. DEBUG, INFO, WARN, ERROR}
}
\value{

}
\description{
Initializes logger for Trajectories package
}
