% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertTable.R
\name{insertTable}
\alias{insertTable}
\title{Replica of insertTable, but to handle schema problem https://github.com/OHDSI/SqlRender/issues/258 for SqlRender}
\usage{
insertTable(connection, databaseSchema = NULL, ...)
}
\description{
Replica of insertTable, but to handle schema problem https://github.com/OHDSI/SqlRender/issues/258 for SqlRender
}
