% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryArguments.R
\name{is.TrajectoryAnalysisArgs}
\alias{is.TrajectoryAnalysisArgs}
\title{Checks whether the object is of type TrajectoryAnalysisArgs}
\usage{
is.TrajectoryAnalysisArgs(x)
}
\arguments{
\item{x}{Any R object}
}
\value{

}
\description{
Checks whether the object is of type TrajectoryAnalysisArgs
}
