% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryArguments.R
\name{is.TrajectoryLocalArgs}
\alias{is.TrajectoryLocalArgs}
\title{Checks whether the object is of type TrajectoryLocalArgs}
\usage{
is.TrajectoryLocalArgs(x)
}
\arguments{
\item{x}{Any R object}
}
\value{

}
\description{
Checks whether the object is of type TrajectoryLocalArgs
}
