% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryArguments.R
\name{IsValidationMode}
\alias{IsValidationMode}
\title{Returns TRUE if event_pairs_for_validation.tsv exists in input folder - means that the package is run in validation mode.}
\usage{
IsValidationMode(trajectoryLocalArgs, verbose = F)
}
\arguments{
\item{trajectoryLocalArgs}{Object created by Trajectories::createTrajectoryLocalArgs() method}

\item{verbose}{If TRUE, outputs some info in INFO/DEBUG log level. Otherwise, returns the results silently.}
}
\value{

}
\description{
Returns TRUE if event_pairs_for_validation.tsv exists in input folder - means that the package is run in validation mode.
}
