% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRenderTranslateSql.R
\name{loadRenderTranslateSql}
\alias{loadRenderTranslateSql}
\title{Copy of SqlRender::loadRenderTranslateSql() method to corrrectly solve pathToSql inside Trajectories package}
\usage{
loadRenderTranslateSql(
  sqlFilename,
  packageName,
  dbms = "sql server",
  ...,
  oracleTempSchema = NULL,
  warnOnMissingParameters = TRUE
)
}
\arguments{
\item{warnOnMissingParameters}{}
}
\value{

}
\description{
Copy of SqlRender::loadRenderTranslateSql() method to corrrectly solve pathToSql inside Trajectories package
}
