\name{over.STTDF.SpatialPolygons}
\alias{over.STTDF.SpatialPolygons}
\title{
spatio-temporal overlap of STTDF and Polygons
}
\description{
spatio-temporal overlap of STTDF and Polygons
}
\usage{
over.STTDF.SpatialPolygons(sttdf, pol)
}

\arguments{
  \item{sttdf}{
  a STTDF object
}
  \item{pol}{
  a SpatialPolygons object
}
}

\value{
\code{results} return a single SpatialPolygons object that contains the trajectories that overlap with the spatial polygon
}
\references{
TBA
}
\author{
Jinlong Yang: jinlong@psu.edu
}