\name{over.STTDF.STTDF}
\alias{over.STTDF.STTDF}

\title{
spatio-temporal overlap of two STTDF objects
}
\description{
spatio-temporal overlap of two STTDF objects
}
\usage{
over.STTDF.STTDF(x, y, returnList = FALSE, ...)
}

\arguments{
  \item{x}{
  a STTDF object
}
  \item{y}{
  a STTDF object
}
  \item{returnList}{
  whether to returnList or not
}
 \item{...}{
  ...
}
}

\value{
\code{super} return a single STTDF object with all attribute values stored in the data slot of STTDF object.
}
\references{
TBA
}
\author{
Jinlong Yang: jinlong@psu.edu
}