\name{plot.arwlen}
\alias{plot.arwlen}
\title{
Methods for class arwlen
}
\description{
Methods for class arwlen.
}
\usage{
\method{plot}{arwlen}(x, ...)
}
\arguments{
  \item{x}{ an object of class  arwlen.}
  \item{...}{arguments passed to plot.}
}
\value{
a plot.
}
\author{
Mohammad Mehdi Moradi <m2.moradi@yahoo.com>
}
\seealso{
\link[trajectories]{avemove}
}