\name{plot.distrack}
\alias{plot.distrack}
\title{
Methods for class distrack
}
\description{
The plot method for class distrack.
}
\usage{
\method{plot}{distrack}(x, ...)
}
\arguments{
\item{x}{ an object of class distrack.}
\item{...}{arguments passed to plot.}
}

\author{
Mohammad Mehdi Moradi <m2.moradi@yahoo.com>
}
