\name{plot}
\alias{plot.STI}
\alias{plot.STTDF}
\title{
Ploting STI and STTDF objects
}
\description{
Plot a map with trajectories from STI or STTDF objects. Trajectories from STTDF objects are coloured by trip
}
\usage{
plot.STI(x, y, ...)
plot.STTDF(x, y, ...)
}

\arguments{
  \item{x}{
  an STI or STTDF object
}
  \item{y}{
  missing
  }
  \item{...}{
  ...
}
}


\author{
Jinlong Yang: jinlong@psu.edu
}