% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIgraph.R
\name{plotIgraph}
\alias{plotIgraph}
\title{Title}
\usage{
plotIgraph(
  g,
  layout = layout_nicely(g),
  outputPdfFullpath = F,
  nodesizes = V(g)$count,
  linknumbers = E(g)$numcohortExact,
  title = ""
)
}
\arguments{
\item{outputPdfFullpath}{}
}
\value{

}
\description{
Title
}
