% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrajectoriesGraph.R
\name{plotTrajectoriesGraph}
\alias{plotTrajectoriesGraph}
\title{Plot TrajectoriesGraph object to PDF file}
\usage{
plotTrajectoriesGraph(
  g,
  layout = layout_nicely(g),
  outputPdfFullpath = F,
  nodesizes = V(g)$count,
  linknumbers = E(g)$numcohortExact,
  title = ""
)
}
\arguments{
\item{g}{TrajectoriesGraph object}

\item{layout}{layout of an igraph object}

\item{outputPdfFullpath}{Full path to output PDF file}

\item{nodesizes}{Which values to use for node sizes. By default, uses count values of the nodes from TrajectoriesGraph object}

\item{linknumbers}{Which numbers to show on edges. By default, uses numcohortExact values of the edges from TrajectoriesGraph object}

\item{title}{Title of the graph}
}
\value{

}
\description{
Plot TrajectoriesGraph object to PDF file
}
