% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTrajectoriesGraphForEvent.R
\name{PlotTrajectoriesGraphForEvent}
\alias{PlotTrajectoriesGraphForEvent}
\title{Plots graph for a specific event}
\usage{
PlotTrajectoriesGraphForEvent(
  connection,
  trajectoryAnalysisArgs,
  trajectoryLocalArgs,
  g,
  eventId,
  limitOfNodes = 50
)
}
\arguments{
\item{limitOfNodes}{}
}
\value{

}
\description{
Plots graph for a specific event
}
