% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTrajectoriesGraphForEvents.R
\name{PlotTrajectoriesGraphForEvents}
\alias{PlotTrajectoriesGraphForEvents}
\title{Creates plots for the analysis results obtained by runEventPairAnalysis()}
\usage{
PlotTrajectoriesGraphForEvents(
  connection,
  trajectoryAnalysisArgs,
  trajectoryLocalArgs,
  eventIds = NA,
  limitOfNodes = 30,
  skipOutputTables = T
)
}
\arguments{
\item{connection}{Database connection object created by createConnectionDetails() method in DatabaseConnector package}

\item{trajectoryAnalysisArgs}{TrajectoryAnalysisArgs object that must be created by createTrajectoryAnalysisArgs() method}

\item{trajectoryLocalArgs}{TrajectoryLocalArgs object that must be created by createTrajectoryLocalArgs() method}

\item{eventIds}{Exact concept id-s as a list that is used for analyzing/plotting trajectories. If not specified (NA) (recommended) creates trajectories for top 5 events.}

\item{limitOfNodes}{Max number of events in the consrtructed graph. Used to limit the analysis to most important events only.}

\item{skipOutputTables}{If set to TRUE, no output data tables are made (the PDF graphs only).}
}
\value{

}
\description{
Analysis results, obtained by runEventPairAnalysis(), have to exist in output folder, set by Trajectories::GetOutputFolder(trajectoryLocalArgs,trajectoryAnalysisArgs).
This function also alignes actual trajectories to the graph, therefore it requires database connection.
If you provide an eventName (OMOP concept name) or list of event names as input, the graph that crosses that event is shown. If you do not provide it (eventName=NA), the function automatically takes
most prevalent top5 events and builds graph for each of them.
}
