\name{print.ArimaTrack}
\alias{print.ArimaTrack}
\title{
Methods for class ArimaTrack
}
\description{
print method.
}
\usage{
\method{print}{ArimaTrack}(x)
}

\arguments{
  \item{x}{ an  object of class ArimaTrack.}
}
\author{
<<<<<<< HEAD
Mohammad Mehdi Moradi <m2.moradi@yahoo.com>
}
=======
Mohammad Mehdi Moradi <moradi@uji.es>
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

>>>>>>> refs/remotes/edzer/master
