\name{print.gTrack}
\alias{print.gTrack}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Methods for class "gTrack"
}
\description{
print method.}
\usage{
print.gTrack(X)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
an object of class "gTrack"}
}
\details{
none.}
\value{
none.}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Mohammad Mehdi Moradi <moradi@uji.es>}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
