\name{print.KTrack}
\alias{print.KTrack}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Methods for class "KTrack"
}
\description{
Methods for class "KTrack"
}
\usage{
\method{print}{KTrack}(x,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ an object of class "KTrack" }
  \item{...}{ ignored }
}
\details{ to print an object of class "KTrack".}
\author{
Mohammad Mehdi Moradi <moradi@uji.es>
}

