\name{print.ppplist}
\alias{print.ppplist}
\title{
Print Methods
}
\description{
  method to print an object of class ppplist}
\usage{
  \method{print}{ppplist}(x,...)
}
\arguments{
  \item{x}{an object of class ppplist.}
  \item{...}{ignored}
}
\author{
Mohammad Mehdi Moradi \email{m2.moradi@yahoo.com}
}

  