\name{print.Track}
\alias{print.Track}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Methods for class "Track"
}
\description{
method to print an object of class "Track"}
\usage{
print.Track(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
an object of class "Track"
}
}
\author{
Mohammad Mehdi Moradi <moradi@uji.es>
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

