\name{print.Tracks}
\alias{print.Tracks}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Methods for class "Tracks"
}
\description{
method to print an object of class "Tracks"}
\usage{
\method{print}{Tracks}(x,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ an object of class "Tracks" }
  \item{...}{ ignored }
}
\author{
Mohammad Mehdi Moradi <moradi@uji.es>
}


