\name{print.Trrow}
\alias{print.Trrow}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Methods for class "Trrow"
}
\description{
Print objetcs of class "Trrow"}
\usage{
print.Trrow(X)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
an object of class "Trrow"
}
}
\details{
none.
}
\value{
none.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Mohammad Mehdi Moradi <moradi@uji.es>
}
\note{
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
as.Track.arrow
}
\examples{
print.Trrow(x)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
