% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeOrphanNodes.R
\name{removeOrphanNodes}
\alias{removeOrphanNodes}
\title{Creates force network graph from the event pairs}
\usage{
removeOrphanNodes(g)
}
\arguments{
\item{g}{igraph object}
}
\value{

}
\description{
Creates force network graph from the event pairs
}
