% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setRole.R
\name{setRole}
\alias{setRole}
\title{This function switches role of the database session by simply executing SET ROLE ...;}
\usage{
setRole(connection, trajectoryLocalArgs)
}
\arguments{
\item{sqlRole}{SQL role to be set}
}
\value{

}
\description{
This function switches role of the database session by simply executing SET ROLE ...;
}
