\name{summary}
\alias{summary.STI}
\alias{summary.STTDF}
\title{
Output the summary of STI and STTDF objects
}
\description{
Outpout the summary of STI and STTDF objects
}
\usage{
summary.STI(object, ...)
summary.STTDF(object, ...)
}

\arguments{
  \item{object}{
  an STI or STTDF object
}
  \item{...}{
  ...
}
}


\author{
Jinlong Yang: jinlong@psu.edu
}