\name{traj_sample}

\title{Sample trajectory data for testing}
\alias{Sample trajectory data for testing}
\description{
Public GPS traces from alex 18 at OpenStreetMap.org/trace/ . Five trajectories are available in this sample as five STI objects stored in a single list
}

\usage{
data(traj_sample)
}
\author{Public trace data obtained by Jinlong Yang from http://www.openstreetmap.org/traces/ . }
\references{ 
http://www.openstreetmap.org/user/alex\%2018/traces/1507959
http://www.openstreetmap.org/user/alex\%2018/traces/1507961
http://www.openstreetmap.org/user/alex\%2018/traces/1507963
http://www.openstreetmap.org/user/alex\%2018/traces/1507965
http://www.openstreetmap.org/user/alex\%2018/traces/1507966
}

\keyword{datasets}
\examples{
data(traj_sample)
}
