\name{traj}
\alias{traj}
\title{ miscelaneous functions }
\description{
read the GEO life trajectories
}
\usage{
	traj_reader(path, start = 0, end = 181)
	toSTTDF(traj)
	speed_cal(sttdf)
	hourly_stats(sttdf)
}
\arguments{
\item{path}{ path}
\item{start}{ start}
\item{end}{end}
\item{traj}{ trajectory}
\item{sttdf}{object of class \link[spacetime]{STTDF}}
}
\value{ t.b.d. }
\examples{
# examples go here

}
\keyword{spatial}
