% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryArguments.R
\name{TrajectoryAnalysisArgsFromInputFolder}
\alias{TrajectoryAnalysisArgsFromInputFolder}
\title{Searches for trajectoryAnalysisArgs.json file from inputFolder (defined in trajectoryLocalArgs), creates trajectoryAnalysisArgs object from it and returns it.}
\usage{
TrajectoryAnalysisArgsFromInputFolder(trajectoryLocalArgs)
}
\arguments{
\item{trajectoryLocalArgs}{Object created by Trajectories::createTrajectoryLocalArgs() method}
}
\value{
TrajectoryLocalArgs object
}
\description{
Searches for trajectoryAnalysisArgs.json file from inputFolder (defined in trajectoryLocalArgs), creates trajectoryAnalysisArgs object from it and returns it.
}
