% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryArguments.R
\name{TrajectoryAnalysisArgsFromJson}
\alias{TrajectoryAnalysisArgsFromJson}
\title{Reads trajectoryAnalysisArgs object from JSON file}
\usage{
TrajectoryAnalysisArgsFromJson(filepath)
}
\arguments{
\item{filepath}{Full path to JSON file}
}
\value{
TrajectoryAnalysisArgs object
}
\description{
Reads trajectoryAnalysisArgs object from JSON file
}
