% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryArguments.R
\name{TrajectoryAnalysisArgsToJson}
\alias{TrajectoryAnalysisArgsToJson}
\title{Writes trajectoryAnalysisArgs object to JSON file}
\usage{
TrajectoryAnalysisArgsToJson(trajectoryAnalysisArgs, filepath)
}
\arguments{
\item{trajectoryAnalysisArgs}{Object created by Trajectories::createTrajectoryAnalysisArgs() method}

\item{filepath}{Full path to the output file. Should have .json extension as this is actually a JSON file.}
}
\value{

}
\description{
Writes trajectoryAnalysisArgs object to JSON file
}
