% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualisingTrajectories.R
\name{visualize_data_pairs}
\alias{visualize_data_pairs}
\title{Visualize data pairs using Shiny application}
\usage{
visualize_data_pairs(data)
}
\arguments{
\item{data}{Data in xlsx format used for visualization. Has to contain columns E1_CONCEPT_ID and E2_CONCEPT_ID}
}
\description{
Visualize data pairs using Shiny application
}
\examples{
data <- as.data.frame(read.xlsx("inst/shiny/VisualisingTrajectories/Data/event_pairs_tested.xlsx"))
Trajectories::visualize_data_pairs(data)
}
