% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBirthsChecker.R
\name{addBirthsChecker}
\alias{addBirthsChecker}
\title{Checks whether person.birth_datetime is recorded in the data if addBirths is set to TRUE}
\usage{
addBirthsChecker(connection, trajectoryAnalysisArgs, trajectoryLocalArgs)
}
\arguments{
\item{connection}{Database connection object created by createConnectionDetails() method in DatabaseConnector package}

\item{trajectoryAnalysisArgs}{TrajectoryAnalysisArgs object that must be created by createTrajectoryAnalysisArgs() method}

\item{trajectoryLocalArgs}{TrajectoryLocalArgs object that must be created by createTrajectoryLocalArgs() method}
}
\value{

}
\description{
Checks whether person.birth_datetime is recorded in the data if addBirths is set to TRUE
}
