% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignActualTrajectoriesToGraph.R
\name{alignActualTrajectoriesToGraph}
\alias{alignActualTrajectoriesToGraph}
\title{Adds numcohortCustom value to graph edges - actual number of people (out of all people who have eventId) on that edge}
\usage{
alignActualTrajectoriesToGraph(
  connection,
  trajectoryAnalysisArgs,
  trajectoryLocalArgs,
  g,
  eventid,
  limit = 1000,
  filename = file.path(getwd(), "trajectories.csv"),
  filename_interpretation = file.path(getwd(), "trajectories_interpretation.txt")
)
}
\arguments{
\item{connection}{Database connection object created by createConnectionDetails() method in DatabaseConnector package}

\item{trajectoryAnalysisArgs}{TrajectoryAnalysisArgs object that must be created by createTrajectoryAnalysisArgs() method}

\item{trajectoryLocalArgs}{TrajectoryLocalArgs object that must be created by createTrajectoryLocalArgs() method}

\item{g}{An igrpah object that is created by specific graph functions in this package}

\item{eventid}{Event name (concept ID) through which the trajectories are analyzed}

\item{limit}{Max number of trajectories to align (to limit the analysis). Set to NA if no limit.}

\item{filename}{Full path to output file for trajectory counts. Set to NA to skip this.}

\item{filename_interpretation}{Full path to output file for textual interpretation of the output graph. Set to NA to skip this.}
}
\value{

}
\description{
Adds numcohortCustom value to graph edges - actual number of people (out of all people who have eventId) on that edge
}
