% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEventPairsTable.R
\name{createEventPairsTable}
\alias{createEventPairsTable}
\title{Creates event pairs table and populates it with the data}
\usage{
createEventPairsTable(connection, trajectoryAnalysisArgs, trajectoryLocalArgs)
}
\arguments{
\item{connection}{Database connection object created by createConnectionDetails() method in DatabaseConnector package}

\item{trajectoryAnalysisArgs}{TrajectoryAnalysisArgs object that must be created by createTrajectoryAnalysisArgs() method}

\item{trajectoryLocalArgs}{TrajectoryLocalArgs object that must be created by createTrajectoryLocalArgs() method}
}
\value{

}
\description{
Creates event pairs table and populates it with the data
}
