% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEventPairsTable.R
\name{createEventPairsTable}
\alias{createEventPairsTable}
\title{Creates event pairs table and populates it with the data}
\usage{
createEventPairsTable(connection, trajectoryAnalysisArgs, trajectoryLocalArgs)
}
\arguments{
\item{connection}{DatabaseConnectorConnection object that is used to connect with database}

\item{trajectoryAnalysisArgs}{Object created by Trajectories::createTrajectoryAnalysisArgs() method}

\item{trajectoryLocalArgs}{Object created by Trajectories::createTrajectoryLocalArgs() method}
}
\value{

}
\description{
Creates event pairs table and populates it with the data
}
