% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFilteredFullgraphs.R
\name{createFilteredFullgraphs}
\alias{createFilteredFullgraphs}
\title{Creates full graph plots from event pairs (not filtered to specific concept id-s)}
\usage{
createFilteredFullgraphs(
  connection,
  trajectoryAnalysisArgs,
  trajectoryLocalArgs
)
}
\arguments{
\item{connection}{Database connection object created by createConnectionDetails() method in DatabaseConnector package}

\item{trajectoryAnalysisArgs}{TrajectoryAnalysisArgs object that must be created by createTrajectoryAnalysisArgs() method}

\item{trajectoryLocalArgs}{TrajectoryLocalArgs object that must be created by createTrajectoryLocalArgs() method}
}
\value{

}
\description{
Creates full graph plots from event pairs (not filtered to specific concept id-s)
}
