% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIgraph.R
\name{createIgraph}
\alias{createIgraph}
\title{Title}
\usage{
createIgraph(
  packageName,
  connection,
  sqlRole,
  resultsSchema,
  prefixForResultTableNames,
  eventPairResultsFilename,
  outputFolder,
  cohortName = "",
  eventName = ""
)
}
\arguments{
\item{outputFolder}{Output folder (should exist) without trailing slash}
}
\value{

}
\description{
Title
}
