% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIgraph.R
\name{createIgraph}
\alias{createIgraph}
\title{Creates igraph plots for the analysis results}
\usage{
createIgraph(
  connection,
  trajectoryAnalysisArgs,
  trajectoryLocalArgs,
  eventName = NA
)
}
\arguments{
\item{connection}{Database connection object created by createConnectionDetails() method in DatabaseConnector package}

\item{trajectoryAnalysisArgs}{TrajectoryAnalysisArgs object that must be created by createTrajectoryAnalysisArgs() method}

\item{trajectoryLocalArgs}{TrajectoryLocalArgs object that must be created by createTrajectoryLocalArgs() method}

\item{eventName}{Exact concept name that is used for building trajectories. Must exist in event pairs data table. If not specified (NA) (recommended) creates trajectories for top 5 events.}
}
\value{

}
\description{
Analysis results have to exist in output folder, set by Trajectories::GetOutputFolder(trajectoryLocalArgs,trajectoryAnalysisArgs).
This function also alignes actual trajectories to the graph, therefore it requires database connection.
}
