% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIgraph.R
\name{createIgraph}
\alias{createIgraph}
\title{Title}
\usage{
createIgraph(
  connection,
  trajectoryAnalysisArgs,
  trajectoryLocalArgs,
  eventPairResultsFilename,
  outputFolder,
  eventName = NA
)
}
\arguments{
\item{connection}{Database connection object created by createConnectionDetails() method in DatabaseConnector package}

\item{trajectoryAnalysisArgs}{TrajectoryAnalysisArgs object that must be created by createTrajectoryAnalysisArgs() method}

\item{trajectoryLocalArgs}{TrajectoryLocalArgs object that must be created by createTrajectoryLocalArgs() method}

\item{eventPairResultsFilename}{File path to event pairs that is used for building an igraph object.}

\item{outputFolder}{Output folder (should exist) without trailing slash}

\item{eventName}{Exact concept name that is used for building trajectories. Must exist in event pairs data table. If not specified (NA) (recommended) creates trajectories for top 5 events.}
}
\value{

}
\description{
Title
}
