% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryArguments.R
\name{createTrajectoryAnalysisArgs}
\alias{createTrajectoryAnalysisArgs}
\title{Title}
\usage{
createTrajectoryAnalysisArgs(
  minimumDaysBetweenEvents,
  maximumDaysBetweenEvents,
  minPatientsPerEventPair,
  addConditions,
  addObservations,
  addProcedures,
  addDrugExposures,
  addDrugEras,
  addBirths,
  addDeaths,
  daysBeforeIndexDate,
  packageName = "Trajectories",
  cohortName
)
}
\arguments{
\item{minimumDaysBetweenEvents}{The smallest number of days between 2 events of the patient that can be considered as event pair. Usually we have used 1.}

\item{maximumDaysBetweenEvents}{The maximum number of days between 2 events of the patient that can be considered as event pair. Ususally we have not really limited it so we have used 3650 (10 years)}

\item{minPatientsPerEventPair}{Minimum number of people having event1 -> event2 progression to be included in analysis. Can be used for limiting analysis to frequent event pairs only. However, it does not throw less frequent diagnosis pairs out of the (control group) data and therefore, does not affect the statistical significance.}

\item{addConditions}{TRUE/FALSE parameter to indicate whether events from Condition_occurrence table should be included in the analysis}

\item{addObservations}{TRUE/FALSE parameter to indicate whether events from Condition_occurrence table should be included in the analysis}

\item{addProcedures}{TRUE/FALSE parameter to indicate whether events from Procedure_occurrence table should be included in the analysis}

\item{addDrugExposures}{TRUE/FALSE parameter to indicate whether events from Drug_exposure table should be included in the analysis}

\item{addDrugEras}{TRUE/FALSE parameter to indicate whether events from Drug_era table should be included in the analysis. NB! use either addDrugEras=T or addDrugExposures=T (not both) as it leads to analysis duplication...}

\item{addBirths}{TRUE/FALSE parameter to indicate whether births events should be included in the analysis.}

\item{addDeaths}{TRUE/FALSE parameter to indicate whether events from Death table should be included in the analysis.}

\item{daysBeforeIndexDate}{0 or any positive number that indicates for how many days before index date of the cohort the events are included in the analysis. In case one wants to include all events before index date, use value Inf}

\item{packageName}{do not edit, this is required by SqlRender::loadRenderTranslateSql}

\item{cohortName}{Reader-friendly short description of the cohort. Used in graph titles and file names (can contain spaces)}
}
\value{

}
\description{
Title
}
