% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryArguments.R
\name{createTrajectoryLocalArgs}
\alias{createTrajectoryLocalArgs}
\title{An object to hold local database-specific parameters}
\usage{
createTrajectoryLocalArgs(
  cdmDatabaseSchema,
  vocabDatabaseSchema,
  resultsSchema,
  oracleTempSchema,
  sqlRole = F,
  prefixForResultTableNames = "",
  cohortTableSchema,
  cohortTable,
  cohortId,
  mainOutputFolder = getwd(),
  databaseHumanReadableName = "My database",
  cohortSqlFile
)
}
\arguments{
\item{cdmDatabaseSchema}{Schema containing source data in OMOP CDM format}

\item{vocabDatabaseSchema}{Schema containing OMOP vocabulary}

\item{resultsSchema}{Schema the user has writing access to (used to write analysis tables into)}

\item{oracleTempSchema}{In case you are using oracle, schema for temporary tables need to be specified. A schema where temp tables can be created in Oracle. Otherwise leave it as it is (is not used)}

\item{sqlRole}{Role to use in SQL for writing tables in 'resultsSchema'. It should also have access to 'cdmDatabaseSchema' and 'vocabDatabaseSchema'. Set to FALSE (or F) if setting to a specific role is not needed. It should be safe to use F if you have no idea of what the SQL roles mean.}

\item{prefixForResultTableNames}{Table prefix that is used for all output tables to avoid any collision with existing table names. An empty string is also allowed.}

\item{cohortTableSchema}{Schema where cohort table is located}

\item{cohortTable}{Name of the cohort table in cohortTableSchema}

\item{cohortId}{ID of the cohort (in cohortTable) that will be used for the analysis}

\item{mainOutputFolder}{The output folder path. This is the folder where the final results are produced into. Use full path and do NOT add trailing slash! The folder must already exist. Default value is the default working directory.}

\item{databaseHumanReadableName}{In the future, it will be added to the titles of the graph to indicate what data is this. Use something short. Currently this parameter is not used.}

\item{cohortSqlFile}{Name of the built-in cohort file to be used when creating cohorts. Currently, you can use 'example_cohort_RA.sql', for instance}
}
\value{

}
\description{
An object to hold local database-specific parameters
}
