% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createValidationSetup.R
\name{createValidationSetup}
\alias{createValidationSetup}
\title{Creates setup file for the validation of the results in a separate database}
\usage{
createValidationSetup(trajectoryAnalysisArgs, trajectoryLocalArgs)
}
\arguments{
\item{trajectoryAnalysisArgs}{TrajectoryAnalysisArgs object that must be created by createTrajectoryAnalysisArgs() method}

\item{trajectoryLocalArgs}{TrajectoryLocalArgs object that must be created by createTrajectoryLocalArgs() method}
}
\value{

}
\description{
Creates setup file for the validation of the results in a separate database
}
