% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbCleanup.R
\name{dbCleanup}
\alias{dbCleanup}
\title{Removes all database tables that were created during the analysis}
\usage{
dbCleanup(connection, trajectoryAnalysisArgs, trajectoryLocalArgs)
}
\arguments{
\item{connection}{DatabaseConnectorConnection object that is used to connect with database}

\item{packageName}{Must always have value 'Trajectories'. The value is needed by SqlRender to find the SQL scripts from the right path.}

\item{dbms}{The target SQL dialect that is used by SqlRender}

\item{oracleTempSchema}{A schema that can be used to create temp tables in when using Oracle. Used by SqlRender}

\item{sqlRole}{Database role that was used when creating/dropping tables in resultsSchema. Set to F if a specicif role is not needed}

\item{resultsSchema}{Database schema where the temporary analysis tables were created.}

\item{prefixForResultTableNames}{This is the prefix that was used for all table names in analysis process. Default value is ''. This function is going to delete tables starting with that prefix.}
}
\value{

}
\description{
Removes all database tables that were created during the analysis
}
