% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbCleanup.R
\name{dbCleanup}
\alias{dbCleanup}
\title{Removes all database tables that were created during the analysis}
\usage{
dbCleanup(connection, trajectoryAnalysisArgs, trajectoryLocalArgs)
}
\arguments{
\item{connection}{Database connection object created by createConnectionDetails() method in DatabaseConnector package}

\item{trajectoryAnalysisArgs}{TrajectoryAnalysisArgs object that must be created by createTrajectoryAnalysisArgs() method}

\item{trajectoryLocalArgs}{TrajectoryLocalArgs object that must be created by createTrajectoryLocalArgs() method}
}
\value{

}
\description{
Removes all database tables that were created during the analysis
}
