% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropCohortTable.R
\name{dropCohortTable}
\alias{dropCohortTable}
\title{Function to drop cohort table.}
\usage{
dropCohortTable(
  packageName,
  connection,
  dbms,
  sqlRole = "",
  cohortTableSchema = cohortTableSchema,
  cohortTable = "my_cohorts"
)
}
\arguments{
\item{connection}{DatabaseConnectorConnection object that is used to connect with database}

\item{sqlRole}{SQL role that is used to drop cohort table in cohortTableSchema. Set to empty string ('') if setting a specific role is not needed.}

\item{cohortTable}{}
}
\value{

}
\description{
Function to drop cohort table.
}
