% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropCohortTableTO_DELETE.R
\name{dropCohortTable}
\alias{dropCohortTable}
\title{Function to drop cohort table.}
\usage{
dropCohortTable(connection, trajectoryAnalysisArgs, trajectoryLocalArgs)
}
\arguments{
\item{connection}{Database connection object created by createConnectionDetails() method in DatabaseConnector package}

\item{trajectoryAnalysisArgs}{TrajectoryAnalysisArgs object that must be created by createTrajectoryAnalysisArgs() method}

\item{trajectoryLocalArgs}{TrajectoryLocalArgs object that must be created by createTrajectoryLocalArgs() method}
}
\value{

}
\description{
Function to drop cohort table.
}
