% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillCohortTable.R
\name{fillCohortTable}
\alias{fillCohortTable}
\title{Function to fill the cohort table.}
\usage{
fillCohortTable(connection, trajectoryAnalysisArgs, trajectoryLocalArgs)
}
\arguments{
\item{connection}{DatabaseConnectorConnection object that is used to connect with database}

\item{cohortSqlFile}{SQL file in the package directory/inst/sql/sql_server). If set to F, uses built-in (example) cohort definition (example_cohort_RA.sql).}

\item{target_cohort_id}{ID that is given to the created cohort. Default is 1.}
}
\value{

}
\description{
Function to fill the cohort table.
}
