% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByLift.R
\name{filterByLift}
\alias{filterByLift}
\title{Filtering edges by lift}
\usage{
filterByLift(g, lift.threshold = 2)
}
\arguments{
\item{g}{igraph object}

\item{lift.threshold}{lift.threshold}
}
\value{

}
\description{
Filtering edges by lift
}
