% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCohortSize.R
\name{getCohortSize}
\alias{getCohortSize}
\title{Gets cohort size (number of rows in cohort table of that particluar cohort ID)}
\usage{
getCohortSize(connection, trajectoryAnalysisArgs, trajectoryLocalArgs)
}
\arguments{
\item{connection}{Database connection object created by createConnectionDetails() method in DatabaseConnector package}

\item{trajectoryAnalysisArgs}{Object created by Trajectories::createTrajectoryAnalysisArgs() method}

\item{trajectoryLocalArgs}{Object created by Trajectories::createTrajectoryLocalArgs() method}
}
\value{

}
\description{
Gets cohort size (number of rows in cohort table of that particluar cohort ID)
}
