% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryArguments.R
\name{GetOutputFolder}
\alias{GetOutputFolder}
\title{Returns full path to output folder for the results.}
\usage{
GetOutputFolder(
  trajectoryLocalArgs,
  trajectoryAnalysisArgs,
  createIfMissing = F
)
}
\arguments{
\item{trajectoryLocalArgs}{Object created by Trajectories::createTrajectoryLocalArgs() method}

\item{trajectoryAnalysisArgs}{Object created by Trajectories::createTrajectoryAnalysisArgs() method}

\item{createIfMissing}{If TRUE, then creates necessary folder if missing.}
}
\value{
Full output path
}
\description{
Basically combines the value of mainOutputFolder, database name, and analysis name to get the output folder. Checks also that the folder exists. If createIfMissing=T, then creates the necessary subfolders under mainOutputFolder.
}
