% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryArguments.R
\name{IsValidationMode}
\alias{IsValidationMode}
\title{Returns TRUE if the package is run in validation mode.}
\usage{
IsValidationMode(trajectoryAnalysisArgs, verbose = F)
}
\arguments{
\item{trajectoryAnalysisArgs}{Object created by Trajectories::createTrajectoryAnalysisArgs() method}

\item{verbose}{If TRUE, outputs some info in INFO/DEBUG log level. Otherwise, returns the results silently.}
}
\value{

}
\description{
Returns TRUE if the package is run in validation mode.
}
