\name{over.STTDF.STTDF}
\alias{over.STTDF.STTDF}

\title{
spatio-temporal overlap of two STTDF objects
}
\description{
spatio-temporal overlap of two STTDF objects
}
\usage{
over.STTDF.STTDF(x, y, returnList = FALSE, ...)
}

\arguments{
  \item{x}{
  a STTDF object
}
  \item{y}{
  a STTDF object
}
  \item{returnList}{
  whether to returnList or not
}
 \item{...}{
  ...
}
}

\value{
\code{super} return a single STTDF object with all attribute values stored in the data slot of STTDF object.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
