\name{plot.arwlen}
\alias{plot.arwlen}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Methods for class "arwlen"
}
\description{
Methods for class "arwlen"
}
\usage{
\method{plot}{arwlen}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ an object of class  "arwlen"}
  \item{...}{ passed on to plot}
}
\value{
a plot.
}
\author{
Mohammad Mehdi Moradi <moradi@uji.es>
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{avemove}
