\name{plot.distrack}
\alias{plot.distrack}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Methods for class "distrack"
}
\description{
The plot method for "distrack" objects.
}
\usage{
plot.distrack(X,xlab="time",ylab="average distance",\dots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{an object of class "distrack"}
  \item{xlab}{the x label of the plot}
  \item{ylab}{the y label of the plot}
  \item{\dots}{further plotting parameters}
}
\details{
This plots an object of class "distrack".
}
\value{
None.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Mohammad Mehdi Moradi <moradi@uji.es>
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
