\name{plot.distrack}
\alias{plot.distrack}
\title{
Plot Methods
}
\description{
The plot method for class distrack.
}
\usage{
\method{plot}{distrack}(x, ...)
}
\arguments{
\item{x}{ an object of class distrack.}
\item{...}{arguments passed to plot.}
}

\author{
Mohammad Mehdi Moradi \email{m2.moradi@yahoo.com}
}
\seealso{
\link[trajectories]{avedistTrack}
}
