\name{plot.distrack}
\alias{plot.distrack}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Methods for class "distrack"
}
\description{
The plot method for "distrack" objects.
}
\usage{
plot(X,xlab="time",ylab="average distance",...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
an object of class "distrack"
}
\item{...}{further plotting parameters}
}
\details{
This plots an object of class "distrack".
}
\value{
None.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Mohammad Mehdi Moradi <moradi@uji.es>
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
