\name{plot.distrack}
\alias{plot.distrack}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Methods for class "distrack"
}
\description{
The plot method for "distrack" objects.
}
\usage{
\method{plot}{distrack}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ an object of class "distrack" }
  \item{...}{ ignored }
}
\details{
This plots an object of class "distrack".
}
\author{
Mohammad Mehdi Moradi <moradi@uji.es>
}
