\name{plot.gTrack}
\alias{plot.gTrack}
\title{
Methods for class gTrack.
}
\description{
plot the pair correlation function for a trajectory pattern.
}
\usage{
\method{plot}{gTrack}(x, type = "l", col = "grey70",cex=1,line=2.2, ...)
}
\arguments{
  \item{x}{ an object of class gTrack.}
  \item{type}{line type.}
  \item{col}{line color.}
  \item{cex}{used for size of legend.}
  \item{line}{specifying a value for line overrides the default placement of labels, see \link[graphics]{title}.}
  \item{...}{arguments passed to plot.}
}
\author{
Mohammad Mehdi Moradi <m2.moradi@yahoo.com>}
