\name{plot.gTrack}
\alias{plot.gTrack}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Methods for class "gTrack"
}
\description{
plot method}
\usage{
\method{plot}{gTrack}(x, type = "l", col = "grey70", ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ an object of class "gTrack" }
  \item{type}{ line type}
  \item{col}{ line color}
  \item{...}{ passed on to plot}
}

\author{
Mohammad Mehdi Moradi <moradi@uji.es>}

