% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTrajectoriesGraphForEvent.R
\name{PlotTrajectoriesGraphForEvent}
\alias{PlotTrajectoriesGraphForEvent}
\title{Plots graph for a specific event}
\usage{
PlotTrajectoriesGraphForEvent(
  connection,
  trajectoryAnalysisArgs,
  trajectoryLocalArgs,
  g,
  eventId,
  limitOfNodes = 30,
  skipOutputTables = T
)
}
\arguments{
\item{skipOutputTables}{If set to T, no output data tables are made (the PDF graphs only).}
}
\value{

}
\description{
Plots graph for a specific event
}
