\name{print.ArimaTrack}
\alias{print.ArimaTrack}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Methods for class "ArimaTrack"
}
\description{
print method.}
\usage{
\method{print}{ArimaTrack}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ an  object of class "ArimaTrack" }
  \item{...}{ignored}
}
\author{
Mohammad Mehdi Moradi <moradi@uji.es>
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
