\name{print.ArimaTrack}
\alias{print.ArimaTrack}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Methods for class "ArimaTrack"
}
\description{
print method.}
\usage{
\method{print}{ArimaTrack}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ an  object of class "ArimaTrack" }
  \item{...}{ignored}
}
\author{
Mohammad Mehdi Moradi <moradi@uji.es>
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

