\name{print.arwlen}
\alias{print.arwlen}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
print.arwlen
}
\description{
to print an object of class "arwlen".}
\usage{
print.arwlen(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
an object of class "arqlen"
}
}
\details{
none.}
\value{
none.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Mohammad Mehdi Moradi <moradi@uji.es>
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
none.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
