\name{print.arwlen}
\alias{print.arwlen}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Methods for class "arwlen"
}
\description{
to print an object of class "arwlen".}
\usage{
print.arwlen(X)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
an object of class "arqlen"
}
}

\author{
Mohammad Mehdi Moradi <moradi@uji.es>
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
