\name{print.arwlen}
\alias{print.arwlen}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Methods for class "arwlen"
}
\description{
to print an object of class "arwlen".}
\usage{
\method{print}{arwlen}(x,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ an object of class "arqlen" }
  \item{...}{ignored}
}
\author{
Mohammad Mehdi Moradi <moradi@uji.es>
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

