\name{print.gTrack}
\alias{print.gTrack}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Methods for class "gTrack"
}
\description{
print method.}
\usage{
print.gTrack(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
an object of class "gTrack"}
}
\author{
Mohammad Mehdi Moradi <moradi@uji.es>}

%% ~Make other sections like Warning with \section{Warning }{....} ~

