\name{print.Track}
\alias{print.Track}
\alias{print.Tracks}
\alias{print.TracksCollection}

\title{
Print Methods
}
\description{
Print method for classes \link[trajectories]{Track}, \link[trajectories]{Tracks}, \link[trajectories]{TracksCollection}.
}
\usage{
\method{print}{Track}(x)
\method{print}{Tracks}(x)
\method{print}{TracksCollection}(x)

}
\arguments{
  \item{x}{ an object of classes \link[trajectories]{Track}, \link[trajectories]{Tracks}, \link[trajectories]{TracksCollection}}
}
\author{
Mohammad Mehdi Moradi \email{m2.moradi@yahoo.com}
}
