\name{print.Tracks}
\alias{print.Tracks}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Methods for class "Tracks"
}
\description{
method to print an object of class "Tracks"}
\usage{
print.Tracks(X)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
an object of class "Tracks"
}
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Mohammad Mehdi Moradi <moradi@uji.es>
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
