\name{print.Tracks}
\alias{print.Tracks}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Methods for class "Tracks"
}
\description{
method to print an object of class "Tracks"}
\usage{
print.Tracks(X)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
an object of class "Tracks"
}
}
\author{
Mohammad Mehdi Moradi <moradi@uji.es>
}


