\name{print.TracksCollection}
\alias{print.TracksCollection}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Methods for class "TracksCollection"
}
\description{
method to print an object of class "TracksCollection"}
\usage{
\method{print}{TracksCollection}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ an object of class "TracksCollection" }
  \item{...}{ ignored }
}
\author{
Mohammad Mehdi Moradi <moradi@uji.es>
}
