\name{print.TracksCollection}
\alias{print.TracksCollection}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Methods for class "TracksCollection"
}
\description{
method to print an object of class "TracksCollection"}
\usage{
print.TracksCollection(X)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
an object of class "TracksCollection"
}
}
\author{
Mohammad Mehdi Moradi <moradi@uji.es>
}
