\name{print.Trrow}
\alias{print.Trrow}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Methods for class "Trrow"
}
\description{
Print objetcs of class "Trrow"}
\usage{
print.Trrow(X)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
an object of class "Trrow"
}
}
\author{
Mohammad Mehdi Moradi <moradi@uji.es>
}

\seealso{
as.Track.arrow
}
