% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeOrphanNodes.R
\name{removeOrphanNodes}
\alias{removeOrphanNodes}
\title{Removes orphan nodes from igraph/TrajectoriesGraph object}
\usage{
removeOrphanNodes(g)
}
\arguments{
\item{g}{igraph object}
}
\value{

}
\description{
Removes orphan nodes from igraph/TrajectoriesGraph object
}
