% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{The main function to run the Trajectory package in validation mode - that is, to validate someone's results in your local OMOP CDM}
\usage{
validate(
  connection,
  trajectoryLocalArgs,
  createCohort = T,
  createEventPairsTable = T,
  runValidationAnalysis = T,
  forceRecalculationOfAnalysis = F,
  createFilteredFullgraphs = T,
  runTrajectoryAnalysis = F,
  cleanup = T
)
}
\arguments{
\item{connection}{Database connection object created by createConnectionDetails() method in DatabaseConnector package}

\item{trajectoryLocalArgs}{TrajectoryLocalArgs object that must be created by createTrajectoryLocalArgs() method}

\item{createCohort}{Builds a study cohort in the database. Normally set to T/TRUE but for debugging or running the analysis step-by-step one can set it to F/FALSE as well.}

\item{createEventPairsTable}{Builds all event pairs and necessary data tables in the database for the analysis. Normally set to T/TRUE but for debugging or running the analysis step-by-step one can set it to F/FALSE as well.}

\item{runValidationAnalysis}{Run the actual directionality analysis of all event pairs. Normally set to T/TRUE but for debugging or running the analysis step-by-step one can set it to F/FALSE as well.}

\item{forceRecalculationOfAnalysis}{Used only when runValidationAnalysis=T. If TRUE, forces deleting previous results from the database and rerunning the whole validation analysis. Useful mostly in case something goes wrong and you need to force the recalculation (it is, when debugging). In normal circumstances using FALSE is safe.}

\item{createFilteredFullgraphs}{Builds graphs based on the results. Normally set to T/TRUE but for debugging or running the analysis step-by-step one can set it to F/FALSE as well.}

\item{runTrajectoryAnalysis}{If TRUE, runs the trajectory analysis - puts the actual trajectories to the graph}

\item{cleanup}{Drops tables from the database that were created during various stages of the analysis. Normally set to T/TRUE but for debugging or running the analysis step-by-step one can set it to F/FALSE as well.}
}
\description{
The main function to run the Trajectory package in validation mode - that is, to validate someone's results in your local OMOP CDM
}
